d/control, d/rules: feature architecture parity
Enable systemtap, numa and numad on architecture having the build-deps
nowadays.

Ubuntu enabled these features on ppc64el and s390x that work fine there and
have no reason to be non-enabled there based on these architectures.

On review it was pointed out that the build-deps are linux-any these
days, therefore enable it less selective on linux-any.

Signed-off-by: Dimitri John Ledkov <xnox@ubuntu.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>