gbp: Import and export upstream signatures
We've stored upstream signatures in the pristine-tar branch for
a while now, but until recently gbp didn't automatically process
them.

Now that #872864 has been fixed, we can finally tell gbp that
upstream produces signed releases and that the signatures should
always be imported and exported along with the corresponding
release archive.

Gbp-Dch: Ignore