rules: Install upstream release notes
The ChangeLog file for libvirt hasn't contained any useful
information since 5.3.0, and installing it is entirely optional
these days; on the other hand, as of Policy 4.2.0 it's suggested
that upstream release notes are installed if available, so let's
ship the latter instead of the former.

We install the release notes directly instead of linking back to
the primary package: unlike the original ChangeLog, the release
notes are pretty small, especially when compressed, so having
multiple copies on the system is not a big problem.