rules: Add dh_auto_install override
Ideally, the upstream build system would install all files exactly
where they need to be for the various dh_* tools to pick them up,
but that is not always possible so some tweaking is necessary.

Right now we're mixing these two steps (preparing files so that
dh_*  will pick them up and actually calling dh_*), but things
become cleaner and easier to reason about when we separate them.

Gbp-Dch: Ignore