rules: Rename upstream polkit rules
While upstream has included this file for a long time, they use
a slightly different name than the one we want to use in Debian
for historical reasons as well as to match the other polkit file
we install. Change its name before dh_install is called to
achieve the desired results.

Gbp-Dch: Ignore