rules: Enable ZFS support correctly
Starting with

  commit 3ce2f8690933b72f65c42ebaa2ccbe8821e1eecd
  Author: Guido Günther <agx@sigxcpu.org>
  Date:   Sun Aug 14 15:28:28 2016 +0200

    Support zfs on Linux too

ZFS support is supposed to be enabled both for Linux and for
non-Linux, which makes it unnecessary to use the WITH_STORAGE_ZFS
variable.

Moreover, while

  commit ee7cd36509d10535628ca402990675644578671b
  Author: Guido Günther <agx@sigxcpu.org>
  Date:   Tue Aug 9 06:58:44 2016 +0200

    Enable zfs pool support

    Closes: #827245

was intended to initially enable ZFS support on non-Linux only,
in fact the logic that was introduced with that commit was along
the lines of

  if linux:
      zfs = False
  else:
      if linux:
          zfs = True

which means that ZFS only ended up being enabled on non-Linux
because the newly-introduced Build-Depends on zfsutils triggered
libvirt's feature autodetection logic.

Gbp-Dch: Ignore