patches: Backport libvirt-guests fix correctly
When the patch was originally backported, it was cherry-picked
to the top of the patch queue, and more specifically on top of

  debian/Debianize-libvirt-guests.patch

which it conflicts with. The result is no longer a straighforward
cherry-pick, which is undesirable, and also merely delays the
need to tweak the Debian-specific patch until we import 6.7.0.

Move the backport to the bottom of the patch queue, before any
Debian-specific changes.

Gbp-Dch: Ignore