gbp: Switch debian-branch to debian/master
We regularly switch between unstable and experimental when
uploading, and that results in a lot of unnecessary churn; we
rarely, if ever, work on more than a single version at any
given time.

Let's instead adopt debian/master, as defined in DEP-14, as
our development branch, and pick the upload target at the last
minute when updating the Debian changelog.

Gbp-Dch: Ignore