rules: Increase timeout for tests to 15m
The recent 6.7.0-1 upload has FTBFS on a few architectures (namely
armel, mips64el, mipsel and riscv64) not due to compiler errors or
anything like that, but because of test suite failures which were
caused by the fact that Meson automatically aborts any test
program that takes longer than 30s to complete.

Since some buildds are known to be quite slow, bump the limit to
15m in order to give them a chance to complete.

Note that this is likely an excessively generous time allocation,
but I currently have no idea what a reasonable one would be: once
builds have completed successfully and we see how long running the
tests actually takes, we can reduce it accordingly.