rules: Make dh_missing errors non-fatal for -indep builds
dh_missing is smart enough to figure out that files that are part
of packages that are not being built, such as -arch packages
during an -indep build, are not to be reported, but of course we
can't expect it to go as far as inspecting debian/rules and
realize that some files have been skipped on purpose.

This hasn't caused any issues so far, but after switching to
debhelper compat 13 with a9cc391d8ec9 dh_missing's default has
changed to --fail-missing, and as a consequence of that -indep
builds are failing. Let's go back to --list-missing for those.