d/control: fix circular dependency on libvirt-daemon-driver-qemu (Closes: #963898)
Back when we split the connection drivers we wanted to make the most
common case 'qemu' to be depended on. But lintian rightfully yells at us

  W: libvirt source: intra-source-package-circular-dependency libvirt-daemon
  libvirt-daemon-driver-qemu

While it is the most common case there is no strict need to use it with
qemu - so a recommends makes more sense here.

The drawback is that people that run install-recommends:false will not
get a working qemu/kvm setup, but that would be correct for what they
asked apt to be installed.

Fixes: 8dc356ce "Move qemu, lxc, uml, vbox and xen connection drivers into
separate packages."

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>