rules: Introduce ARCHES_* variables
Some features (namely the lxc, xen and vbox drivers) can only be
enabled on a subset of all possible Linux architectures, and this
fact is reflected both when installing build dependencies and when
choosing which arguments to pass to Meson.

Using explicit architecture lists makes it easier to make sure
they're always up to date and to quickly verify that the data is
consistent between debian/rules and debian/control.

This is similar in spirit to upstream commit

  https://gitlab.com/libvirt/libvirt/-/commit/974dc0a4c678af8ccb9224abecc834bb593e81fa

Gbp-Dch: Ignore