libvirt-clients: Move out virt-qemu-run
This tool requires the qemu driver to be available locally in
order to work, so it belongs in the libvirt-daemon-driver-qemu
package.

People who only had libvirt-clients installed before upgrading
would not have been able to use it successfully, and for those
who had a setup where virt-qemu-run worked things will be
shuffled around smoothly during upgrade, so there's no need to
make a big deal about this change.