libvirt-daemon: Move out sanlock-related files
The libvirt_sanlock_helper and virt-sanlock-cleanup utilities
are for exclusive use of the sanlock driver, and as such they
belong with it rather than with the daemon itself.

The transition will be completely transparent to anyone with a
working sanlock installation, so no need to make a fuss about
it.