libvirt-clients: Add Recommends for libvirt-login-shell
If this is not present, then apt will not know that the package
needs to be installed, and virt-login-shell will disappear from
the user's system during upgrade, which is something that we
don't want.

We will probably want to weaken this to a Suggests for bookworm,
as most people don't need virt-login-shell to be installed on
their machines, but not breaking existing setups is more
important and Recommends still makes it possible to remove the
package after upgrade, so it's a good compromise.

Gbp-Dch: Ignore