libvirt-login-shell: Transfer conffile from libvirt-clients correctly
If we simply remove the conffile from libvirt-clients and add it
to libvirt-login-shell, at the end of the update any local
modification that was made in the past will only be present in
the .dpkg-bak file, with the actual conffile being reverted back
to its default state.

Since dpkg-maintscript-helper doesn't currently have support for
tranferring conffiles between packages, we have to implement the
logic ourselves.

Gbp-Dch: Ignore