libvirt-daemon-config-network: New binary package
Currently, the libvirt-daemon-system package contains the
configuration file for the 'default' libvirt network; however,
while this is convenient to have in many scenarios, it might be
unwanted in others: the current solution for this scenario is
to install the package and manually remove the conffile
afterwards.

Moreover, part of the configuration is autogenerated the first
time the daemon reads the file, and this results in annoying
dpkg prompts on package upgrade.

In order to address both issues, the configuration file is
moved to a separate binary package, and is no longer marked as
a conffile but custom handling is added to the maintainer
scripts instead: this way, not having the default network
becomes just a matter of not installing the additional package,
and the dpkg prompts are no longer triggered on upgrade.

Closes: #973489