rules: Move virt-aa-helper to libvirt-daemon
It's an executable part of the daemon rather than configuration,
the latter being what the libvirt-daemon-system is supposed to
contain.

It just happens to be the last bit of architecture-dependent
content left in the package too, so with that out of the way we
can consider making it Arch: all in the future.