libvirt-daemon-config-*: Install target directories
Although the dependency on libvirt-daemon-system currently
ensure that these directories already exist when the packages
are installed, that's no longer going to be the case once we
can drop the strong Depends and turn it into a weaker
Recommends.

Additionally, including the directory in the packages will
serve as a way for the user to figure out where the
configuration that's generated from the included templates is
stored, so it'd be nice to have regardless of future changes.

Gbp-Dch: Ignore