control: Strengthen dependencies between packages
When splitting a package, such as we did when we moved the QEMU
driver from libvirt-daemon to libvirt-daemon-driver-qemu, you
need to maintain at least some level of dependency between the
original package and the new one, or apt will simply not know
that it needs to install the latter in an upgrade scenario.

Using Recommends works for most cases, since Debian systems are
configured to install those by default; however, there are a
couple of cases where it's better to have a stronger dependency.

More specifically, due to the way libvirt has been packaged
until now, people expect that installing the libvirt-daemon
package will result in the QEMU driver being available, and
that installing the libvirt-daemon-system package will result
in the default network and default set of nwfilters being
configured.

The stronger dependencies help both for upgrades and in the
context of other packages' autopkgtests; unfortunately, it
also makes the package less flexible. We're going turn these
Depends back into Recommends once bullseye is out.