control: Avoid circular dependencies
Circular dependencies are best avoided because they can cause
dependency resolution issues, especially during upgrades. In
order to break the cycle, make it so the more specific package,
such as libvirt-daemon-driver-qemu, no longer depends on the
more generic package, in this case libvirt-daemon.

Unfortunately, this means that the user might end up installing
a virtualization driver without getting the daemon itself
installed, which is not the best experience. We will be able to
improve upon that in bookworm by reintroducing these
dependencies once bullseye is out.