systemd: start, but do not restart libvirt-guests.service
Due to various changes libvirt-guests.service is no more started at all.
This all started when trying to not restart it on upgrades which we
now do properly for LIBVIRT_SYSTEM_SERVICES, so move libvirt-guests.service
to this group.

Not doing so would stop and start VMs on package upgrades.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>