d/control: avoid libvirt-clients to pull in libvirt-daemon
As of today `libvirt-clients` pulls in too much due to the
split of `libvirt-login-shell` and the dependencies it has now.
Not everyone using `libvirt-clients` (I'd think only a very small
fraction) uses `libvirt-login-shell`, but it is a Recommends and therefore
often installed as dependency.

Further down `libvirt-login-shell` depends on `libvirt-daemon-driver-lxc`
(reasonable), but than in turn depends on `libvirt-daemon`.
That latter one is reasonable for `libvirt-daemon-driver-lxc` itself, but
looking at the full chain it means that anyone using `libvirt-clients` (which
could quite often be to manage a remote system) will also get `libvirt-daemon`
installed.

Cut this dependency chain by reducing `libvirt-clients` -> `libvirt-login-shell`
to a suggests.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>