rules: Install default files in a simpler way
Currently, we're installing the upstream-provided default files
in a rather roundabout way: we copy them from the source
directory to the debian/ directory, then we ask dh_installinit
to to install them, and we clean up the copy at dh_auto_clean
time.

The upstream build system already installs all these files
under $(DESTDIR), and all we need to do is rename the directory
they are installed into so that it matches Debian conventions
and pick up the files from there.

Gbp-Dch: Ignore