libvirtd: Patch upstream default file
Right now we're ignoring the upstream-provided default file and
installing our own that we keep in debian/ instead, but that
means that a) we might not realize upstream has made some
changes that our default file should adapt to and b) we have to
treat this one default file differently from the other ones we
install.

Since we can't just install the upstream default file directly,
settle for the next best thing: tweak it via a patch before the
upstream build system installs it.

For now, the patch replaces the upstream file completely with
our version, but later we're going to aim for a more
maintainable hybrid approach.

Gbp-Dch: Ignore