libvirtd: Improve default file
Several changes:

  * use $LIBVIRTD_ARGS, which is consistent with upstream and
    other daemons, instead of $libvirtd_opts;

  * drop mentions of $KRB5_KTNAME, same as upstream has done
    with commit 68d08cf2f200;

  * list all sysv-only options together, and make it very clear
    that they don't apply when systemd is in use.

Note that, unlike upstream, we don't enable a timeout for
libvirtd by default: this is because we need the default file
to work both for systemd, which has socket activation, and for
sysv, which doesn't.

During upgrades, if the default file had not been modified by
the user then the transition will happen quietly; however, if
local modifications had been made, then the admin will be left
dealing with a nasty-looking diff: highlighting the change in
a NEWS entry will hopefully mitigate the confusion.