sysv: Refactor sysv init scripts
There are a few cases in which hardcoded values are used when a
suitable variable already exists; some conditionals in the script
for libvirtd are also tweaked so that a diff between the two
scripts will only highlight the actual functional differences,
without additional noise.

This commit is better viewed with 'git show -w'.

Gbp-Dch: Ignore