libvirt-daemon-system-sysv: Add Lintian overrides
Starting with 2b705f6e5c95, the libvirt-guests sysv init script
is just a wrapper, and the actual functionality is implemented
by a script installed under /usr/lib/libvirt which is common to
both sysv and systemd.

Lintian is, understandably, not aware of this, and thus flags
the sysv init script as containing a number of serious mistakes.
Add overrides to silence it.

Gbp-Dch: Ignore