rules: Disable netcf support
As of 7.1.0, the spec file included upstream disables this.

While the Meson configuration will still enable the feature if
the necessary header files are found, overall netcf support (and
the netcf project itself) is basically considered a failed
experiment at this point, and the expectation is that usage will
naturally disappear over time.

https://listman.redhat.com/archives/libvir-list/2020-December/msg00183.html