rules: Explicitly set remote_default_mode
This knob controls whether the client library will attempt to
autostart the modular or monolithic daemon when a connection
attempt is made and no daemon is running.

The old behavior is "legacy" (spawn the monolithic daemon),
whereas upstream has switched to "direct" (spawn the modular
daemons) as the default starting with libvirt 7.5.0.

Since we don't yet ship the modular daemons, we need to make
sure we use "legacy" mode instead of the upstream default.