libvirt-daemon-system: Make QEMU cache directory root-owned
In the past, this directory was used for things like domain
screenshot data, and so it needed to be writable by the user
under which the QEMU process runs.

As of libvirt 7.9.0, those files are now written in per-domain
directories and all that remains in this directory is the
capability cache for the QEMU driver, so the more open
permissions are no longer necessary.