salsa: Use standard pipeline
Back when we enabled CI support in libvirt, Salsa would hang
because of the massive amount of output produced during a
build of libvirt.

Either the switch to Meson has improved the situation or Salsa
has been changed in ways that make it less affected by it:
either way, the standard pipeline works fine now. We just need
to disable the reprotest job because libvirt is currently not
reproducible.

Gbp-Dch: Ignore