control: Improve multiarch support
The changes made to -doc, -daemon-system-systemd and -wireshark
are those suggested by the multiarch hinter. The hinter also
suggests to make the various drivers Multi-Arch: same, but that
can't work because they all depend on -daemon, which is not
co-installable. Hopefully making them explicitly Multi-Arch: no
will make the hinter stop suggesting those changes; even if
not, there's no harm in spelling out the status quo.