control: Drop all Pre-Depends
The ones in libvirt0 and libnss-libvirt were added when
introducing multiarch support, and the current implementation
no longer requires them so they always end up being empty.

The ones in libvirt-daemon-* expand to

  Pre-Depends: init-system-helpers (>= 1.54~)

which is guaranteed to be satisfied on any of the reasonable
backport targets, namely Debian 11 bullseye and Ubuntu 20.04
LTS focal.