Implement custom handling for systemd units
Take the systemd unit handling code we're currently using, which
is a slightly tweaked version of the code dh_installsystemd would
generate, and change it further so that it becomes easier to
understand and maintain.

Since the underlying logic is ultimately extremly similar, the
new code should work as expected both when installing the
package from scratch and when upgrading from a version where
dh_installsystemd (or code derived from it) was still in use.