control: Fix cross building
Starting with version 1:2.6.1-1, nfs-common Depends on python3.

As a result, when cross building and targeting for example
arm64, nfs-common:arm64 will be installed to satisfy libvirt's
Build-Depends and python3:arm64 will in turn be installed to
satisfy nfs-common's Depends.

Attempting to do so will however result in an error:

  Setting up python3.10-minimal:arm64 (3.10.5-1) ...
  /var/lib/dpkg/info/python3.10-minimal.postinst: 51: /usr/bin/python3.10: Exec format error
  dpkg: error processing package python3.10-minimal:arm64 (--configure):
   installed python3.10-minimal:arm64 package post-installation script subprocess returned error exit status 126
  Errors were encountered while processing:
   python3.10-minimal:arm64
  E: Sub-process /usr/bin/dpkg returned an error code (1)
  E: Failed to process build dependencies

In order to avoid this issue, force the native version of
nfs-common to be installed. This was always the version we
wanted anyway.