control: Force more Build-Depends to be native
When cross building, we currently get the foreign version of
these packages instead of the native one.

While that's not affecting the build process negatively,
because we don't attempt to execute any of the binary files
but just check to see whether they exist in $PATH, it's still
more correct to require the native version instead.

Gbp-Dch: Ignore