daemon-system: Don't create README.Debian symlink
It was added in

  commit ec4de6a5b5290955653c2ae0012286ed60ce06e1
  Author: Guido Günther <agx@sigxcpu.org>
  Date:   Wed Sep 2 17:52:36 2015 +0200

    Move README.Debian to libvirt-daemon

    libvirt-bin is a transitional package.

as part of turning libvirt-bin into a transitional package;
that package was later dropped in

  commit 961c93de026997d7d21a88faed08548c25cbc14d
  Author: Guido Günther <agx@sigxcpu.org>
  Date:   Thu Feb 4 09:05:54 2016 +0100

    Drop libvirt-bin transitional package

With recent versions of lintian, the file being a symlink
results in

  Warning in processable .../libvirt-daemon-system_8.4.0-1_amd64.deb: Cannot open
  .../libvirt-daemon-system_8.4.0-1_amd64_binary/unpacked/usr/share/doc/libvirt-daemon-system/README.Debian.gz
  at /usr/share/lintian/lib/Lintian/Check/Debian/Readme.pm line 109.
  warning: cannot run debian/readme check on package binary:libvirt-daemon-system_8.4.0-1_amd64
  skipping check of binary:libvirt-daemon-system_8.4.0-1_amd64

which is an immediate failure. At this point the transitional
package has been gone for years and we don't have any other
cross-package symlinks, so do away with this one too and make
lintian happy again.

Gbp-Dch: Ignore