not-installed: Don't install virt-qemu-qmp-proxy yet
This new tool is written in Python and calls the libvirt API,
which means it requires the python3-libvirt package to be
installed. However, the libvirt-clients package needs to remain
lean, so we don't want it to start depending on Python.

Upstream has introduced the libvirt-client-qemu package to
solve this conundrum, and we should do something like that in
Debian as well.

But the libvirt package in Debian is already a few months out
of date at this point, and introducing a new binary package
would mean a trip through the NEW queue, so let's not install
the new tool yet.

Gbp-Dch: Ignore