libvirt-clients-qemu: New binary package
This contains a couple of QEMU-specific tools.

Besides the fact that they're only intended to work against a
single hypervisor driver, another big motivation for them to
be in a separate binary package is that they, unlike existing
tools such as virsh, are written in Python and as such require
both Python and the Python bindings for libvirt to be
installed in order to work.