libnss-libvirt: Update apt seccomp filter
Installing libnss-libvirt and enabling it in nsswitch.conf causes the
apt http method to issue getdents and getdents64 syscalls, which are not
whitelisted in its seccomp filter. This makes apt unusable for the most
part, and it breaks unattended-upgrades in particular.

/etc/apt/apt.conf.d/90libnss-libvirt already creates an exception for
getdents syscalls. Update that file in order to allow getdents64
syscalls as well.

Closes: #934474

Suggested-by: Thomas Luzat <thomas@luzat.com>
Signed-off-by: Smits Katze <smitsohu@gmail.com>