libvirt-daemon-system: Depend on polkitd instead of policykit-1
The polkit package has been made more modular, and libvirt only
needs the daemon part.

Note that we explicitly ask for support for the legacy pkla
format to be present because, while Debian now contains a version
of polkit recent enough to support the JavaScript-based rules
that upstream includes, Ubuntu still doesn't.

We also don't include a fallback dependency on policykit-1
because, with Debian 12 just a few months away, backports to
Debian 11 don't seem like something that we should concern
ourselves with. Ubuntu 22.04, which we still consider a valid
backport target, implements the polkit package split.

Closes: #1025578