rules: avoid $(findstring) for architecture filtering
$(findstring) will look for matches anywhere in the string,
even within words: this means that it may match unwanted
architectures, e.g. the old architecture "mips" would match
in a list of architectures containing "mipsel".

To avoid this problem, use $(filter) which looks for entire
words in the search text.

A couple of occurrences (the "linux" cases) are switched
directly to string comparison, as it is more logic for them.

Gbp-Dch: Ignore