links: Link /usr/share/doc/* to /usr/share/doc/libvirt0
Instead of needlessly shipping the same files over and over
again, just ship them once as part of the libvirt0 package and
have all the other packages point to that.

This approach is explicitly allowed by Policy.

Ideally we'd implement this using dh_installdocs' --link-doc
option, but the tool refuses to create symlinks between
arch:any and arch:all packages. My understanding is that
having an arch:any -> arch:all symlink would cause issues
with binNUMs, but our case (arch:all -> arch:any) should be
fine.

The only package for which we don't do this is libvirt-doc,
since we ship a bunch of additional documentation files in
that one and so it needs to have its own /usr/share/doc
directory.