control: Make all packages depend on libvirt0
This is already the case for basically all packges, except
it happens indirectly so Lintian can't figure out that's the
case. Making the dependency explicit solves that issue.

libvirt0 is an arch:any package, which poses a challenge when
it comes making it a dependency of arch:all packages.

In general, we want all libvirt packages installed on a
system to have the same version, and that's exactly why we
already have a bunch of '(= ${binary:Version})' type
dependencies. See #981435 for additional information.

However, since binNMUs are a thing, we could end up in a
situation where ${binary:Version} is increased without the
arch:all packages being rebuilt.

This is why we need to use, for arch:all packages, the
slightly more cumbersome version constraints that are based
off ${source:Version} instead of ${binary:Version}. See the
not-binnmuable-all-depends-any Lintian tag for more
information.

Gbp-Dch: Ignore