rules: Introduce ARCHES_QEMU
The QEMU package only builds the system emulators on certain
architectures, with some being explicitly left out.

For those (currently alpha, hppa, ia64, m68k and sh4) it
doesn't make sense for us to try and build the QEMU driver,
as the system emulators are a requirement for it to be of
any use at all.

This should make libvirt buildable on those architectures,
or at the very least get us closer to that goal.