postint: No longer create empty overrides for abstractions
The previous paths were Debian-specific and are now deprecated
in favor of an upstream-compatible solution.

Note that we don't create empty overrides for the new locations:
now that we require AppArmor 3, we can safely use the "include
if exists" directive and thus those placeholders are no longer
needed. Upstream AppArmor's preference is for them *not* to be
created to avoid unnecessary clutter.

We also delete the old overrides if there are no local
customizations. Cleaning up after ourselves, at purge time, is
something that we should have done even before now, in the same
way dh_apparmor already does for the empty overrides that it
creates for profiles.

Gbp-Dch: Ignore