dirs: Don't ship channels directory
Starting with libvirt 9.7.0, channels are created under
/run/libvirt instead of /var/lib/libvirt, so we no longer need
to ship any directory as part of the package.

The migration between the two locations happens seamlessly:
domains that are running when the upgrade is performed keep
using the old directory, while domains that are started
afterwards use the new one.

As a result of that, if there are any running domains during
the upgrade, dpkg will refuse to delete the old directories.
Add a postint snippet that will take care of cleaning up
during a future upgrade, once all domains have been restarted
and are thus using the new location.