postrm: Clean up channels on purge
Channels are runtime data that gets created on domain startup
and deleted on domain shutdown. When we're purging the package,
there's no need to keep them around.

To account for the unlikely scenario in which the user might
have stored valuable data in the directory, we don't force
removal of the directory and instead only delete it if it's
empty once the libvirt-owned files have been removed.

Gbp-Dch: Ignore