rules: Temporarily disable nbdkit support
The situation with nbdkit/libnbd availability is a bit
complicated: it's not available on all architectures that
libvirt currently builds on, but the maintainers haven't
explicitly excluded any of those architectures, so in at least
some cases the setback might be just a temporary one.

Since nbdkit usage in libvirt is not currently permitted by the
AppArmor policy, we want to keep it disabled regardless.