rules: Conditionally enable dependency features
These features were never explicitly enabled or disabled,
so we've always relied on detection for them. It's better
to be explicit.

In all cases, we want the feature to be enabled on Linux
and disabled everywhere else.

Notably, the features that we're enabling/disabling here
are all dependencies of other features that we already
handle explicitly. It really doesn't make any sense for
the upstream build system to expose most of these knobs
separately, but for whatever reason they are.

The corresponding Build-Depends can be made Linux-only.

Note that we are no longer building virt-host-validate,
part of the libvirt-clients package, on non-Linux
architectures. This might seem like a mistake at first,
but it is in fact entirely correct.

The virt-host-validate tool, as the name suggests, is
intended to validate whether the host it's running on
it virtualization-capable, and as such it should be
part of the libvirt-daemon package; the fact that it
currently lives in libvirt-clients instead is merely a
packaging bug.

The same issue affected upstream until somewhat
recently (libvirt 7.3.0) and addressing it in Debian
is a job for another day.

Gbp-Dch: Ignore