rules: Conditionally enable existing features
Currently, several features are explicitly enabled based
on the assumption that we always want them, and that is
certainly true when building on Linux.

However, as we work towards making it possible to build a
minimal, client-only version of libvirt on non-Linux
architectures, we have to revisit this assumption and
switch to enabling most of these features conditionally.

The corresponding Build-Depends can be made Linux-only.

Gbp-Dch: Ignore