scripts: Add scaffolding for processing templates
The newly-introduced process-templates.py script, which is going
to be invoked both manually by maintainers and automatically as
part of each package build, will allow us to stop having to keep
in sync the values of the various ARCH_* debian/rules variables
with the contents of debian/control.

Moreover, it implements a subset of the features of dh-exec,
which will allow us to get rid of most of the explicit dh_install
calls in debian/rules.

Gbp-Dch: Ignore