control: Introduce daemon-common package
As of today, Debian only ships the libvirtd monolithic daemon.

Modular daemons were introduced several years ago as a more
secure and flexible alternative, and they're currently the
recommended deployment option both upstream and in other distros
such as Fedora.

Our long-term goal is to make a modular deployment possible in
Debian too, but a significant amount of work is still necessary
for that to happen. As a preparatory step, move some files that
are needed both for monolithic and modular deployments into a
separate package.

Appropriate package relationship are declared to ensure that
things will behave as expected at upgrade time. The usual dance
is performed to correctly transfer conffiles between packages.

It's worth noting that a couple of commands (virt-admin,
virt-host-validate) are moved into this new package from the
libvirt-clients package. While this might look surprising, it's
actually the right thing to do: both tools are only useful
when run directly on the virtualization host, so they don't
belong in the libvirt-clients package, which can be installed
a client machine to manage a remote hypervisor.

This commit moves files and conffiles, but a few additional
items need to be handled too. That will happen in the next
couple of commits.