templates: Move from debian/templates/ to debian/
Turns out that years of habit when it comes to editing files
in the debian/ directory are hard to undo :)

Keep the templates in the same directory as everything else.

For *.install, the output files are only generated at build
time so there is no ambiguity and you hardly even notice the
difference compared to before.

For control, the two files being side by side is still not a
big problem, because the suffix can be used to tell the
template apart, and I've noticed that getting used to editing
that one instead of the output file took hardly any time.

Gbp-Dch: Ignore