install: Turn all into templates
This makes for a more consistent experience when editing
them, and avoids the churn that would happen in the future
if architecture-dependent files were moved between package,
or were added to packages that didn't have any until then.

The script that processes these templates is fairly fast, so
processing a dozen more doesn't cause a noticeable increase
in build time.

Gbp-Dch: Ignore