dirs: Don't list /usr/share/polkit-1/rules.d
We install a file inside it, which automatically takes care of
creating the directory. No need to list it explicitly.

This was probably needed in the past since polkit support is
conditionally enabled, but these days the binary package itself
is Linux only so we can keep things simple.

Gbp-Dch: Ignore