scripts: Make it possible to turn maintscripts into templates
Maintainer scripts are supposed to be completely stand-alone,
so it's not possible to create a "library" of common code.

We still need to perform the same operations in the context of
several binary packages though, so code sharing is still
happening - it just takes the crude form of copy and paste.

Since we are already processing templates at build time,
however, we can extend the existing tooling to work on
maintainer scripts too.

The idea is fairly simple: we can have a single file that
contains various shell snippets, each delimited by specific
markers and assigned a unique name; then, in each maintainer
script, we can inject the snippets that we need, avoiding
duplication.

Gbp-Dch: Ignore