control: Downgrade several Depends to Recommends
When introducing all the new packages that (partially) replace
existing ones, we have liberally added Depends from the old
packages to the new ones, in order to ensure that existing
installations would not lose functionality on upgrade.

In some cases, however, these strong dependency are undesirable
and can even cause dependency loops, which is something that we
want to avoid.

In order to make things better, we now downgrade several of
those Depends to Recommends.

This still achieves our goal of ensuring smooth upgrades, since
disabling Recommends is a "you get to keep both pieces" kind of
deal, especially during upgrades, while at the same time making
it possible to obtain smaller/cleaner installs by strategically
removing certain packages.

The detailed list of Depends that are weakened to Recommends is
as such:

  daemon-system -> daemon-driver-lxc, daemon-driver-xen
  daemon -> daemon-*
  daemon-plugin-lockd -> daemon
  daemon-plugin-sanlock -> daemon

There is a single scenario in which we don't simply downgrade
the dependency, but drop it altogether instead: that's the

  clients -> daemon-common

dependency. In this case, we really don't want the client
package to drag in part of the daemon package, even if it's
just a few files, since it's very common to install the client
package to connect to a remote hypervisor.

This is safe to do in this specific case because the files
that we've moved between packages, namely the virt-admin and
virt-host-validate commands, only work when run locally on the
hypervisor host, so we can be sure that anyone who actually
needs them also has the full daemon package installed.

Gbp-Dch: Ignore