control: Shuffle external dependencies
We've already taken care of all internal dependencies, now it's
time for the external ones.

Most of the changes are obvious: for example, it makes perfect
sense that the QEMU driver would be the one depending on QEMU.

In a few cases, such as for adduser, dependencies had to be
duplicated since the maintainer script snippet that uses them
has been split across multiple binary packages.

Gbp-Dch: Ignore