rules: Configure firewall backend explicitly
Upstream has just introduced an nftables-based backend and made
it the default. We almost certainly want it to be the default
for trixie, but it's probably a good idea to be conservative in
the short term and give it a little time to stabilize upstream
before making the switch ourselves.

Gbp-Dch: Ignore