snippets: Introduce firstver for conffile transfer
Our logic currently assumes that the conffile we're trying to
transfer has existed in the package since forever, and that
assumption applies to most cases so it hasn't caused any trouble
until now.

However, libvirt 10.4.0 introduced a new conffile that we want
to be able to transfer between packages, and we need to make
sure that its absence is not misinterpreted.

In order to do that, introduce a new parameter "firstver" for
all conffile transfer functions. This will contain the first
version that added the conffile to "pkgfrom".

For all existing conffiles, use 6.0.0-0 as "firstver". This
might look odd, but it ensures that upgrades from the version
of libvirt shipped with Ubuntu 20.04 will work correctly.

We could hunt for more accurate versioning information, but
that would require an annoying amount of additional work and
we don't really care about anything older than that anyway.

Gbp-Dch: Ignore